requirements.txt
discord.py
Pillow
requests
json
PyNaCl
ffmpeg-python
youtube_dl